requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
datasets==2.14.5
dill==0.3.7
docopt==0.6.2
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.17.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.3
pyarrow==13.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
seaborn>=0.12.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tabulate
tokenizers==0.13.3
torch==2.1.0
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
onnxruntime==1.16.1
pandarallel==1.6.5
psutil==5.9.5
pynvml==11.5.0
scikit_learn==1.3.1
scipy
wikipedia==1.4.0
UpSetPlot==0.8.0